ci: rename to .hound.yml (#4708)
As of 1ea53b4, a hound.yml file was added to disable JavaScript linting
with jshint. This is the behavior we want, as the browser follows its
own linting style that isn't JSHint. The correct name for the file is
.hound.yml, so nothing happened.

This also removes .jshintrc, because we don't use jshint to lint.